requirements.txt
requests
PyYAML
pydantic
pytz
pandas
plotly
rich
streamlit
click
sqlite-utils
tenacity
feedparser
rapidfuzz
tldextract
geopy
pydantic[email]
python-dateutil
structlog
httpx
pytest
ruff
